+++ src/pages/index.ts
@@ -5,11 +5,11 @@ export { default as SignupRole } from './init/SignupRole';
 export { default as Signup } from './init/Signup';
 
 // call
-export { default as Call } from './Call';
+export { default as Call } from './call/Call';
 export { default as Search } from './Search';
-export { default as RecordDetail } from './RecordDetail';
-export { default as CallWaiting } from './CallWaiting';
-export { default as CallActive } from './CallActive';
+export { default as RecordDetail } from './call/RecordDetail';
+export { default as CallWaiting } from './call/CallWaiting';
+export { default as CallActive } from './call/CallActive';
 
 // report
 export { default as Report } from './Report';
